Add the ability to export YAML (and make default)
This commit incorporates the `YAML::Tiny` package from CPAN.  I debated
copying it into the respository versus adding it to `requires` in the
`Build.PL` script.  I went with the former because:

1. `YAML::Tiny` is designed to be a single file package.  So it's
lightweight and portable.

2. `YAML::Tiny` hasn't changed since 2013.  So it's not likely to change
in the future.  If it proves to be broken, we can import or require
a larger YAML module.

The plugin now exports a YAML file of the main data structure.  No
analysis is done yet.