Make skipping on reporting absentees the default behavior
We set the option `noms.useall` to be false at object creation.  This
means no reporting on students in the names file but not in the
association database.  Seems like a good idea, since the goal of the
plugin is an analysis of student reponses, and if they don't respond
there's nothing to analyze.

In the process, the `pre_process` subroutine was moved inside of
`analyze`.  This means `export` calls `analyze`, which in turn calls
`pre_process`, rather than `export` calling `pre_process` and `analyze`
in succession.  The result is that `analyze` is a fully functional
external method and doesn't require manually calling `pre_process`
first.  That cleans up things a bit for testing purposes.

Fixes #11.