Test for correctly computed weights (Closes #2)
Two cases of incorrectly computed problem weights have been resolved.
The first was from a question that had to be changed after the fact
from MC to MS.  In doing so I accidentally let the NOTA option open, but
of course no student picked it.  So rescoring eliminated that problem.

The second was from two cases on a single MS problem where a student
selected only one of two correct answers (and no other answers).  The
weight for those students on that one answer was the full question score
since omitting it left the question blank.  This is probably the right
answer to the question.

To work around the fact that the answer weights by student are not
totally constant, we compute the exam answer's weight with the median
instead of the mode.