Visualize FR items with box-and-whisker plots instead
Now there are separate tables for the fixed response and free response
items.  The fixed response items have a bar chart of each answer.  The
free response items do a box-and-whisker plot of the item score.

The plot needs some improvements.  For instance, a score of 0 is likely
an outlier.  So lower and upper whiskers should be moved in.  The
technical term for this is Tukey box plot.  See
[Wikipedia](https://en.wikipedia.org/wiki/Box_plot#Types).

Refer to issue #9.