Detect FR problems better
Instead of matching the title with a regular expression, we see if the
list is a range consisting all of all integers from zero to the maximum.

This will succeed in all my use cases.  If someone makes a free-response
problem which skips some values, or has non-integer values, this will fail.

See Issue #9.  In fact, I'm probably not going to work more on this
until after version 1.0.