Improve autodetection of open questions
Open (free response) questions are detected by looking at the set of
weights.  If they form a range 0..n, the problem is flagged as free
reponse.

But if the weights included two digit numbers (e.g., 0-10), they were
being sorted incorrectly.  This cause the free-response detection to
fail.

The fix was to use perl's integer comparison operator when sorting
the list of weights.

Fixes #24.  See also #9.